Added legal message and changed stuff

+++ fivem-restart/Program.cs
@@ -5,6 +5,24 @@ using System.IO;
 
 namespace fivem_restart
 {
+    /*
+    TinyNvidiaUpdateChecker - Check for NVIDIA GPU driver updates!
+    Copyright (C) 2018 SCRP team
+
+    This program Is free software: you can redistribute it And/Or modify
+    it under the terms Of the GNU General Public License As published by
+    the Free Software Foundation, either version 3 Of the License, Or
+    (at your option) any later version.
+
+    This program Is distributed In the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty Of
+    MERCHANTABILITY Or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License For more details.
+    
+    You should have received a copy Of the GNU General Public License
+    along with this program.  If Not, see <http://www.gnu.org/licenses/>.
+    */
+
     class Program
     {
         public static string currentPath = Directory.GetCurrentDirectory();
@@ -13,31 +31,33 @@ namespace fivem_restart
         static void Main(string[] args)
         {
             Console.Title = "FiveM server";
-            Console.WriteLine("FiveM FXServer launcher developed by the SCRP team");
-            Console.WriteLine("v1.0.0");
+            Console.WriteLine("fivem-restart v1.0.0");
+
+            if (!File.Exists("FXServer.exe")) {
+                Console.WriteLine("FXServer was not found, we cannot run the program! Press any key to exit the program.");
+                Console.ReadKey();
+            }
 
             CheckGit();
 
             if (File.Exists(logFile)) {
+                Console.WriteLine("fivem-restart: removed old log file");
                 File.Delete(logFile);
             }
 
-            // LaunchServer();
-
-            Console.ReadKey();
+            LaunchServer();
         }
 
         private static void LaunchServer()
         {
-            //fxserver.exe
-            //@"+set citizen_dir " + currentPath + @"\citizen\ +exec server.cfg"
+            Console.WriteLine("fivem-restart: starting fxserver . . .");
+            LaunchProcess("FXServer.exe", @"+set citizen_dir " + currentPath + @"\citizen\ +exec server.cfg");
         }
 
         private static void CheckGit()
         {
-            Console.WriteLine("Fetching all scripts from git . . .");
-            foreach (string path in Directory.GetFiles(currentPath + @"\resources", "README.md", SearchOption.AllDirectories))
-            {
+            Console.WriteLine("fivem-restart: fetching all scripts from git");
+            foreach (string path in Directory.GetFiles(currentPath + @"\resources", "README.md", SearchOption.AllDirectories)) {
                 PullProject(path);
             }
         }
@@ -59,17 +79,17 @@ namespace fivem_restart
 
         private static void PullProject(string directoryPath)
         {
-
             try {
                 using (var repo = new Repository(Path.GetDirectoryName(directoryPath))) {
                     PullOptions options = new PullOptions();
                     options.FetchOptions = new FetchOptions();
 
-                    Commands.Pull(repo, new Signature(new Identity("SCRP0", ""), new DateTimeOffset(DateTime.Now)), options);
+                    Commands.Pull(repo, new Signature(new Identity("SCRP0", "scrpbot@gmail.com"), new DateTimeOffset(DateTime.Now)), options);
                 }
             } catch (RepositoryNotFoundException) {
                 //Console.WriteLine();
             } catch (Exception ex) {
+                Console.WriteLine(directoryPath);
                 Console.WriteLine(ex.ToString());
             }
         }
